Refactored the DFT to use f64 and iterators

+++ image_proc/tests/fourier_transform_tests/dft_1d_tests.rs
@@ -1,7 +1,7 @@
 use image_proc::modifications::frequency_domain::fourier_transform::dft_1d;
 use num::Complex;
 
-const ACCURACY: f32 = 1e-6;
+const ACCURACY: f64 = 1e-6;
 
 macro_rules! assert_delta {
     ($x:expr, $y:expr, $d:expr) => {
@@ -11,7 +11,7 @@ macro_rules! assert_delta {
     };
 }
 
-fn test_dft_1d(input: &[f32], expected: &[(f32, f32)]) {
+fn test_dft_1d(input: &[f64], expected: &[(f64, f64)]) {
     let expected: Vec<_> = expected
         .iter()
         .map(|(re, im)| Complex::new(*re, *im))
